requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
werkzeug==3.0.1
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib-bin==19.24.2
numpy==1.24.3
pillow==10.1.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
scipy==1.11.4
